Fix referential integrity violation when deleting

+++ dscommercev2/src/main/java/com/ggocodelab/dscommercev2/services/ProductService.java
@@ -1,14 +1,17 @@
 package com.ggocodelab.dscommercev2.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.ggocodelab.dscommercev2.dto.ProductDTO;
 import com.ggocodelab.dscommercev2.entities.Product;
 import com.ggocodelab.dscommercev2.repositories.ProductRepository;
+import com.ggocodelab.dscommercev2.services.exceptions.DatabaseException;
 import com.ggocodelab.dscommercev2.services.exceptions.ResourceNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
@@ -58,8 +61,7 @@ public class ProductService {
 		} 
 		catch (EntityNotFoundException e) {
 			throw new ResourceNotFoundException("Recurso não encontrado");
-		}
-		
+		}		
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -69,9 +71,22 @@ public class ProductService {
 		entity.setImgUrl(dto.getImgUrl());		
 	}
 	
-	@Transactional
+//	@Transactional
+//	public void delete(Long id) {
+//		repository.deleteById(id);
+//	}
+	
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);			
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}		
 	}
